+++ src/LaravelEnumshare.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace Olivermbs\LaravelEnumshare;
-
-class LaravelEnumshare {}