+++ tests/Pest.php
@@ -1,5 +1,5 @@
 <?php
 
-use Olivermbs\LaravelEnumshare\Tests\TestCase;
+use Olivermbs\Enumshare\Tests\TestCase;
 
 uses(TestCase::class)->in(__DIR__);