+++ tests/TestCase.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Olivermbs\LaravelEnumshare\Tests;
+namespace Olivermbs\Enumshare\Tests;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
-use Olivermbs\LaravelEnumshare\LaravelEnumshareServiceProvider;
+use Olivermbs\Enumshare\EnumshareServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 
 class TestCase extends Orchestra
@@ -13,14 +13,14 @@ class TestCase extends Orchestra
         parent::setUp();
 
         Factory::guessFactoryNamesUsing(
-            fn (string $modelName) => 'Olivermbs\\LaravelEnumshare\\Database\\Factories\\'.class_basename($modelName).'Factory'
+            fn (string $modelName) => 'Olivermbs\\Enumshare\\Database\\Factories\\'.class_basename($modelName).'Factory'
         );
     }
 
     protected function getPackageProviders($app)
     {
         return [
-            LaravelEnumshareServiceProvider::class,
+            EnumshareServiceProvider::class,
         ];
     }